Fixes to returns
Controller functions now return much more valid responses in case of errors (bad requests, not found, etc)

+++ src/main/java/com/akshitbaunthy/restapiassignmentsolution/controller/EmployeeDeleteController.java
@@ -1,26 +1,42 @@
 package com.akshitbaunthy.restapiassignmentsolution.controller;
 
+import com.akshitbaunthy.restapiassignmentsolution.entity.Employee;
 import com.akshitbaunthy.restapiassignmentsolution.service.EmployeeDeleteService;
+import com.akshitbaunthy.restapiassignmentsolution.service.EmployeeReadService;
+import com.akshitbaunthy.restapiassignmentsolution.service.EmployeeUpdateService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/employee/delete")
 public class EmployeeDeleteController {
 
     EmployeeDeleteService employeeDeleteService;
+    EmployeeReadService employeeReadService;
 
     @Autowired
-    private EmployeeDeleteController(EmployeeDeleteService employeeDeleteService) {
+    private EmployeeDeleteController(EmployeeDeleteService employeeDeleteService, EmployeeReadService employeeReadService) {
         this.employeeDeleteService = employeeDeleteService;
+        this.employeeReadService = employeeReadService;
     }
 
-    @GetMapping("/single")
-    public void deleteById(@RequestParam Integer id) {
-        employeeDeleteService.deleteById(id);
+    @DeleteMapping("/single")
+    public ResponseEntity<HttpStatus> deleteById(@RequestParam Integer id) {
+        if (id != 0) {
+            Optional<Employee> fetchedEmployee = employeeReadService.getById(id);
+            if(fetchedEmployee.isPresent()) {
+                employeeDeleteService.deleteById(id);
+                return null;
+            }
+            else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
-
 }
